pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "devops-cli"
version = "0.1.0"
description = "AI-powered Kubernetes resource generator"
authors = [
  { name="DevOps Team", email="devops@example.com" },
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "typer[all]>=0.9.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "ruamel.yaml>=0.17.0",
    "rich>=13.0.0",
]

[project.scripts]
devops-cli = "devops_cli.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["devops_cli*"]
